data_generation/synthetic_ui.py
"""Synthetic UI generation for large-scale dataset creation.

This module generates procedurally created GUI examples with automatic annotation.
Supports multiple platforms and styles:
    - Web UI: HTML/CSS rendering in headless browser
    - Desktop: Qt/GTK/WPF simulated layouts
    - Mobile: Android/iOS simulated layouts
    - Games: In-game UI patterns

Generated data includes:
    - Images: Rendered UI screenshots
    - Annotations: Bounding boxes with element types
    - Metadata: Platform, style, complexity, etc.

Usage:
    from data_generation.synthetic_ui import UIGenerator

    generator = UIGenerator(style='material', platform='web')
    for i in range(10000):
        image, annotations = generator.generate()
        # Save image and annotations

References:
    - Procedural generation: https://en.wikipedia.org/wiki/Procedural_generation
    - Domain randomization: https://arxiv.org/abs/1703.06907
"""

import argparse
from typing import Any

from PIL import Image


class UILayoutGenerator:
    """Procedural UI layout generator.

    Creates random but realistic UI layouts using layout algorithms:
        - Vertical/horizontal stacking
        - Grid layouts
        - Flex box (web-like)
        - Nested hierarchies

    Args:
        width: Canvas width
        height: Canvas height
        platform: Target platform ('web', 'mobile', 'desktop', 'game')
        complexity: Layout complexity ('simple', 'medium', 'complex')
    """

    def __init__(
        self,
        width: int = 1024,
        height: int = 768,
        platform: str = "web",
        complexity: str = "medium",
    ) -> None:
        """Initialize layout generator."""
        super().__init__()
        # TODO: Implement layout generator initialization
        # - Store canvas dimensions
        # - Initialize layout algorithms
        # - Load component templates

    def generate(self) -> dict[str, Any]:
        """Generate random layout.

        Returns:
            Dictionary with:
                - elements: List of element definitions (type, position, size)
                - hierarchy: Nesting structure
                - metadata: Layout metadata
        """
        # TODO: Implement layout generation
        # - Choose layout algorithm
        # - Recursively place components
        # - Add text content
        # - Return layout description
        pass


class UIStyler:
    """Apply visual styling to UI layouts.

    Handles:
        - Color scheme selection
        - Typography (fonts, sizes)
        - Spacing and padding
        - Borders, shadows, effects
        - Material Design, iOS, Windows styles

    Args:
        style: Visual style ('material', 'ios', 'windows', 'flat', 'minimal')
        theme: Color theme ('light', 'dark', 'custom')
    """

    def __init__(
        self,
        style: str = "material",
        theme: str = "light",
    ) -> None:
        """Initialize styler."""
        super().__init__()
        # TODO: Implement styler initialization
        # - Load style definitions
        # - Initialize color schemes
        # - Setup typography

    def apply_style(self, layout: dict[str, Any]) -> dict[str, Any]:
        """Apply styling to layout.

        Args:
            layout: Layout description from UILayoutGenerator

        Returns:
            Styled layout with colors, fonts, effects, etc.
        """
        # TODO: Implement style application
        # - Apply color scheme
        # - Add typography
        # - Add effects (shadows, borders, etc.)
        # - Add textures and backgrounds
        pass

    def randomize(self, layout: dict[str, Any]) -> dict[str, Any]:
        """Apply domain randomization to styled layout.

        Args:
            layout: Styled layout

        Returns:
            Randomized layout
        """
        # TODO: Implement domain randomization
        # - Random color variations
        # - Font size variations
        # - Spacing variations
        # - Element size variations
        pass


class UIRenderer:
    """Render UI layouts to images.

    Converts layout descriptions to pixel images using:
        - PIL/Pillow for simple graphics
        - Selenium/headless browsers for web UIs
        - Game engine bindings (optional)

    Args:
        renderer_type: Rendering backend ('pil', 'browser', 'game')
    """

    def __init__(self, renderer_type: str = "pil") -> None:
        """Initialize renderer."""
        super().__init__()
        # TODO: Implement renderer initialization
        # - Setup rendering backend
        # - Initialize resources (fonts, textures)
        # - Load rendering configuration

    def render(
        self,
        layout: dict[str, Any],
    ) -> Image.Image:
        """Render layout to image.

        Args:
            layout: Styled layout description

        Returns:
            PIL Image
        """
        # TODO: Implement rendering
        # - Create canvas
        # - Draw elements recursively
        # - Apply effects
        # - Return PIL Image
        pass


class AnnotationExtractor:
    """Extract annotations from layout descriptions.

    Produces ground truth annotations:
        - Bounding boxes for each element
        - Element type labels
        - Hierarchical relationships
        - Text content (optional)

    Args:
        num_classes: Number of element type classes
    """

    def __init__(self, num_classes: int = 10) -> None:
        """Initialize annotation extractor."""
        super().__init__()
        # TODO: Implement annotation extractor initialization
        # - Load element class mappings
        # - Setup annotation format

    def extract(self, layout: dict[str, Any]) -> dict[str, Any]:
        """Extract annotations from layout.

        Args:
            layout: Layout description

        Returns:
            Dictionary with:
                - boxes: Bounding boxes (x, y, w, h) for each element
                - labels: Class labels for each element
                - confidences: Annotation confidence scores
                - hierarchy: Element relationships
        """
        # TODO: Implement annotation extraction
        # - Traverse layout tree
        # - Extract positions and sizes
        # - Map to class indices
        # - Return annotation dict
        pass


class UIGenerator:
    """Complete UI generation pipeline.

    Combines layout generation, styling, rendering, and annotation
    to produce dataset samples.

    Args:
        width: Image width
        height: Image height
        platform: Target platform
        style: Visual style
        num_classes: Number of element classes
    """

    def __init__(
        self,
        width: int = 1024,
        height: int = 768,
        platform: str = "web",
        style: str = "material",
        num_classes: int = 10,
    ) -> None:
        """Initialize UI generator."""
        super().__init__()
        # TODO: Implement generator initialization
        # - Create layout generator
        # - Create styler
        # - Create renderer
        # - Create annotation extractor

    def generate(self) -> tuple[Image.Image, dict[str, Any]]:
        """Generate one UI example with annotations.

        Returns:
            - image: PIL Image of rendered UI
            - annotations: Ground truth annotations
        """
        # TODO: Implement generation pipeline
        # - Generate layout
        # - Apply style
        # - Randomize
        # - Render to image
        # - Extract annotations
        # - Return image and annotations
        pass

    def generate_batch(
        self,
        batch_size: int = 32,
    ) -> tuple[list[Image.Image], list[dict[str, Any]]]:
        """Generate batch of UI examples.

        Args:
            batch_size: Number of examples to generate

        Returns:
            - images: List of PIL Images
            - annotations: List of annotation dicts
        """
        # TODO: Implement batch generation
        # - Generate multiple examples
        # - Return lists of images and annotations
        pass


def generate_dataset(
    output_dir: str,
    num_samples: int = 10000,
    platform: str = "web",
    style: str = "material",
    batch_size: int = 32,
) -> None:
    """Generate complete dataset of synthetic UIs.

    Args:
        output_dir: Output directory for dataset
        num_samples: Total number of samples to generate
        platform: Target platform
        style: Visual style
        batch_size: Generation batch size
    """
    # TODO: Implement dataset generation
    # - Create output directory
    # - Create generator
    # - Generate samples in batches
    # - Save images and annotations
    # - Create dataset metadata
    pass


def create_parser() -> argparse.ArgumentParser:
    """Create argument parser for data generation.

    Returns:
        ArgumentParser
    """
    # TODO: Implement argument parser
    # - Output directory
    # - Number of samples
    # - Platform selection
    # - Style selection
    # - Batch size
    # - Seed for reproducibility
    pass


def main(args: argparse.Namespace) -> None:
    """Main data generation entry point.

    Args:
        args: Parsed command-line arguments
    """
    # TODO: Implement main function
    # - Create output directory
    # - Call generate_dataset with arguments
    # - Print statistics
    pass


if __name__ == "__main__":
    args = create_parser().parse_args()
    main(args)
